add md5 hashing for sync file comparison
this change was granted because of platform compatibility issue. Windows prefers to rewrite the last modification date of incoming android files, therefore, hashing was implemented to provide further checks in determining what files are to be synced. With that, creating tempfiles was necessary as well, because hashlib library can't directly access the android filesystem.